refactor: measurements.py
