+++ src/orquestra/quantum/symbolic_simulator.py
@@ -8,7 +8,8 @@ from sympy import Symbol
 from orquestra.quantum.api.backend import QuantumSimulator, StateVector
 from orquestra.quantum.circuits import Circuit, Operation
 from orquestra.quantum.circuits.layouts import CircuitConnectivity
-from orquestra.quantum.measurements import Measurements, sample_from_wavefunction
+from orquestra.quantum.measurements import Measurements
+from orquestra.quantum.wavefunction import sample_from_wavefunction
 
 
 class SymbolicSimulator(QuantumSimulator):
