+++ tests/orquestra/quantum/measurements/measurements_test.py
@@ -11,35 +11,17 @@ import pytest
 
 from orquestra.quantum.distributions import MeasurementOutcomeDistribution
 from orquestra.quantum.measurements import (
-    ExpectationValues,
     Measurements,
-    Parities,
     _check_sample_elimination,
-    check_parity,
-    check_parity_of_vector,
-    concatenate_expectation_values,
     convert_bitstring_to_int,
-    expectation_values_to_real,
     get_expectation_value_from_frequencies,
-    get_expectation_values_from_parities,
-    get_parities_from_measurements,
-    load_expectation_values,
-    load_parities,
-    load_wavefunction,
-    sample_from_wavefunction,
-    save_expectation_values,
-    save_parities,
-    save_wavefunction,
 )
 from orquestra.quantum.openfermion.ops import IsingOperator
-from orquestra.quantum.testing import create_random_wavefunction
 from orquestra.quantum.utils import (
     RNDSEED,
-    convert_bitstrings_to_tuples,
     convert_tuples_to_bitstrings,
     get_ordered_list_of_bitstrings,
 )
-from orquestra.quantum.wavefunction import Wavefunction
 
 
 def remove_file_if_exists(filename):
@@ -49,220 +31,9 @@ def remove_file_if_exists(filename):
         pass
 
 
-def test_expectation_values_io():
-    expectation_values = np.array([0.0, 0.0, -1.0])
-    correlations = []
-    correlations.append(np.array([[1.0, -1.0], [-1.0, 1.0]]))
-    correlations.append(np.array([[1.0]]))
-
-    estimator_covariances = []
-    estimator_covariances.append(np.array([[0.1, -0.1], [-0.1, 0.1]]))
-    estimator_covariances.append(np.array([[0.1]]))
-
-    expectation_values_object = ExpectationValues(
-        expectation_values, correlations, estimator_covariances
-    )
-
-    save_expectation_values(expectation_values_object, "expectation_values.json")
-    expectation_values_object_loaded = load_expectation_values(
-        "expectation_values.json"
-    )
-
-    assert np.allclose(
-        expectation_values_object.values,
-        expectation_values_object_loaded.values,
-    )
-    assert len(expectation_values_object.correlations) == len(
-        expectation_values_object_loaded.correlations
-    )
-    assert len(expectation_values_object.estimator_covariances) == len(
-        expectation_values_object_loaded.estimator_covariances
-    )
-    for i in range(len(expectation_values_object.correlations)):
-        assert np.allclose(
-            expectation_values_object.correlations[i],
-            expectation_values_object_loaded.correlations[i],
-        )
-    for i in range(len(expectation_values_object.estimator_covariances)):
-        assert np.allclose(
-            expectation_values_object.estimator_covariances[i],
-            expectation_values_object_loaded.estimator_covariances[i],
-        )
-
-    remove_file_if_exists("expectation_values.json")
-
-
-def test_real_wavefunction_io():
-    wf = Wavefunction([0, 1, 0, 0, 0, 0, 0, 0])
-    save_wavefunction(wf, "wavefunction.json")
-    loaded_wf = load_wavefunction("wavefunction.json")
-    assert np.allclose(wf.amplitudes, loaded_wf.amplitudes)
-    remove_file_if_exists("wavefunction.json")
-
-
-def test_imag_wavefunction_io():
-    wf = Wavefunction([0, 1j, 0, 0, 0, 0, 0, 0])
-    save_wavefunction(wf, "wavefunction.json")
-    loaded_wf = load_wavefunction("wavefunction.json")
-    assert np.allclose(wf.amplitudes, loaded_wf.amplitudes)
-    remove_file_if_exists("wavefunction.json")
-
-
-def test_sample_from_wavefunction():
-    wavefunction = create_random_wavefunction(4, seed=RNDSEED)
-
-    samples = sample_from_wavefunction(wavefunction, 10000, seed=RNDSEED)
-    sampled_dict = Counter(samples)
-
-    sampled_probabilities = []
-    for num in range(len(wavefunction)):
-        bitstring = format(num, "b")
-        while len(bitstring) < wavefunction.n_qubits:
-            bitstring = "0" + bitstring
-        measurement = convert_bitstrings_to_tuples([bitstring])[0]
-        sampled_probabilities.append(sampled_dict[measurement] / 10000)
-
-    probabilities = wavefunction.get_probabilities()
-    for sampled_prob, exact_prob in zip(sampled_probabilities, probabilities):
-        assert np.allclose(sampled_prob, exact_prob, atol=0.01)
-
-
-def test_sample_from_wavefunction_column_vector():
-    n_qubits = 4
-    expected_bitstring = (0, 0, 0, 1)
-    amplitudes = np.array([0] * (2**n_qubits)).reshape(2**n_qubits, 1)
-    amplitudes[1] = 1  # |0001> will be measured in all cases.
-    wavefunction = Wavefunction(amplitudes)
-    sample = set(sample_from_wavefunction(wavefunction, 500))
-    assert len(sample) == 1
-    assert sample.pop() == expected_bitstring
-
-
-def test_sample_from_wavefunction_row_vector():
-    n_qubits = 4
-    expected_bitstring = (0, 0, 0, 1)
-    amplitudes = np.array([0] * (2**n_qubits))
-    amplitudes[1] = 1  # |0001> will be measured in all cases.
-    wavefunction = Wavefunction(amplitudes)
-    sample = set(sample_from_wavefunction(wavefunction, 500))
-    assert len(sample) == 1
-    assert sample.pop() == expected_bitstring
-
-
-def test_sample_from_wavefunction_list():
-    n_qubits = 4
-    expected_bitstring = (0, 0, 0, 1)
-    amplitudes = [0] * (2**n_qubits)
-    amplitudes[1] = 1  # |0001> will be measured in all cases.
-    wavefunction = Wavefunction(amplitudes)
-    sample = set(sample_from_wavefunction(wavefunction, 500))
-    assert len(sample) == 1
-    assert sample.pop() == expected_bitstring
-
-
-@pytest.mark.parametrize("n_samples", [-1, 0])
-def test_sample_from_wavefunction_fails_for_invalid_n_samples(n_samples):
-    n_qubits = 4
-    amplitudes = [0] * (2**n_qubits)
-    amplitudes[1] = 1
-    wavefunction = Wavefunction(amplitudes)
-    with pytest.raises(ValueError):
-        sample_from_wavefunction(wavefunction, n_samples)
-
-
-def test_parities_io():
-    measurements = [(1, 0), (1, 0), (0, 1), (0, 0)]
-    op = IsingOperator("[Z0] + [Z1] + [Z0 Z1]")
-    parities = get_parities_from_measurements(measurements, op)
-    save_parities(parities, "parities.json")
-    loaded_parities = load_parities("parities.json")
-    assert np.allclose(parities.values, loaded_parities.values)
-    assert len(parities.correlations) == len(loaded_parities.correlations)
-    for i in range(len(parities.correlations)):
-        assert np.allclose(parities.correlations[i], loaded_parities.correlations[i])
-    remove_file_if_exists("parities.json")
-
-
-def test_get_expectation_values_from_parities():
-    parities = Parities(values=np.array([[18, 50], [120, 113], [75, 26]]))
-    expectation_values = get_expectation_values_from_parities(parities)
-
-    assert len(expectation_values.values) == 3
-    assert np.isclose(expectation_values.values[0], -0.47058823529411764)
-    assert np.isclose(expectation_values.values[1], 0.030042918454935622)
-    assert np.isclose(expectation_values.values[2], 0.48514851485148514)
-
-    assert len(expectation_values.estimator_covariances) == 3
-    assert np.allclose(
-        expectation_values.estimator_covariances[0],
-        np.array([[0.014705882352941176]]),
-    )
-    assert np.allclose(
-        expectation_values.estimator_covariances[1], np.array([[0.00428797]])
-    )
-
-    assert np.allclose(
-        expectation_values.estimator_covariances[2], np.array([[0.0075706]])
-    )
-
-
-def test_expectation_values_to_real():
-    # Given
-    expectation_values = ExpectationValues(np.array([0.0 + 0.1j, 0.0 + 1e-10j, -1.0]))
-    target_expectation_values = ExpectationValues(np.array([0.0, 0.0, -1.0]))
-
-    # When
-    real_expectation_values = expectation_values_to_real(expectation_values)
-
-    # Then
-    for value in expectation_values.values:
-        assert not isinstance(value, complex)
-    np.testing.assert_array_equal(
-        real_expectation_values.values, target_expectation_values.values
-    )
-
-
 def test_convert_bitstring_to_int():
     bitstring = (0, 1, 0, 1, 0, 1)
     assert convert_bitstring_to_int(bitstring) == 42
-
-
-def test_check_parity_odd_string():
-    bitstring = "01001"
-    marked_qubits = (1, 2, 3)
-    assert not check_parity(bitstring, marked_qubits)
-
-
-def test_check_parity_even_string():
-    bitstring = "01101"
-    marked_qubits = (1, 2, 3)
-    assert check_parity(bitstring, marked_qubits)
-
-
-def test_check_parity_odd_tuple():
-    bitstring = (0, 1, 0, 0, 1)
-    marked_qubits = (1, 2, 3)
-    assert not check_parity(bitstring, marked_qubits)
-
-
-def test_check_parity_even_tuple():
-    bitstring = (0, 1, 1, 0, 1)
-    marked_qubits = (1, 2, 3)
-    assert check_parity(bitstring, marked_qubits)
-
-
-def test_check_parity_of_vector():
-    bitstrings = np.array([[0, 1, 1, 0, 1], [0, 1, 0, 0, 1]])
-    marked_qubits = (1, 2, 3)
-    assert np.allclose(
-        check_parity_of_vector(bitstrings, marked_qubits), np.array([1, 0])
-    )
-
-
-def test_check_parity_of_vector_with_no_marked_qubits():
-    bitstring = np.array([[1, 0], [0, 0]])
-    marked_qubits = []
-    assert all(check_parity_of_vector(bitstring, marked_qubits))
 
 
 def test_get_expectation_value_from_frequencies():
@@ -273,51 +44,6 @@ def test_get_expectation_value_from_frequencies():
         get_expectation_value_from_frequencies(marked_qubits, bitstring_frequencies),
         -0.5,
     )
-
-
-def test_concatenate_expectation_values():
-    expectation_values_set = [
-        ExpectationValues(np.array([1.0, 2.0])),
-        ExpectationValues(np.array([3.0, 4.0])),
-    ]
-
-    combined_expectation_values = concatenate_expectation_values(expectation_values_set)
-    assert combined_expectation_values.correlations is None
-    assert combined_expectation_values.estimator_covariances is None
-    assert np.allclose(combined_expectation_values.values, [1.0, 2.0, 3.0, 4.0])
-
-
-def test_concatenate_expectation_values_with_cov_and_corr():
-    expectation_values_set = [
-        ExpectationValues(
-            np.array([1.0, 2.0]),
-            estimator_covariances=[np.array([[0.1, 0.2], [0.3, 0.4]])],
-            correlations=[np.array([[-0.1, -0.2], [-0.3, -0.4]])],
-        ),
-        ExpectationValues(
-            np.array([3.0, 4.0]),
-            estimator_covariances=[np.array([[0.1]]), np.array([[0.2]])],
-            correlations=[np.array([[-0.1]]), np.array([[-0.2]])],
-        ),
-    ]
-    combined_expectation_values = concatenate_expectation_values(expectation_values_set)
-    assert len(combined_expectation_values.estimator_covariances) == 3
-    assert np.allclose(
-        combined_expectation_values.estimator_covariances[0],
-        [[0.1, 0.2], [0.3, 0.4]],
-    )
-    assert np.allclose(combined_expectation_values.estimator_covariances[1], [[0.1]])
-    assert np.allclose(combined_expectation_values.estimator_covariances[2], [[0.2]])
-
-    assert len(combined_expectation_values.correlations) == 3
-    assert np.allclose(
-        combined_expectation_values.correlations[0],
-        [[-0.1, -0.2], [-0.3, -0.4]],
-    )
-    assert np.allclose(combined_expectation_values.correlations[1], [[-0.1]])
-    assert np.allclose(combined_expectation_values.correlations[2], [[-0.2]])
-
-    assert np.allclose(combined_expectation_values.values, [1.0, 2.0, 3.0, 4.0])
 
 
 class TestMeasurements:
